SD2: Move world_map_outland to separate file
It's too big at this point
