+++ src/game/AI/ScriptDevAI/scripts/outland/netherstorm.cpp
@@ -38,7 +38,7 @@ EndContentData */
 #include "AI/ScriptDevAI/include/precompiled.h"
 #include "AI/ScriptDevAI/base/escort_ai.h"
 #include "AI/ScriptDevAI/base/pet_ai.h"
-#include "AI/ScriptDevAI/scripts/world/world_map_scripts.h"
+#include "AI/ScriptDevAI/scripts/outland/world_outland.h"
 
 /*######
 ## npc_manaforge_control_console
