+++ src/game/AI/ScriptDevAI/system/ScriptLoader.cpp
@@ -348,6 +348,7 @@ extern void AddSC_shadowmoon_valley();
 extern void AddSC_shattrath_city();
 extern void AddSC_terokkar_forest();
 extern void AddSC_boss_terokk();
+extern void AddSC_OutlandWorldScript();
 extern void AddSC_zangarmarsh();
 
 void AddScripts()
@@ -696,5 +697,6 @@ void AddScripts()
     AddSC_shattrath_city();
     AddSC_terokkar_forest();
     AddSC_boss_terokk();
+    AddSC_OutlandWorldScript();
     AddSC_zangarmarsh();
 }
